requirements.txt
streamlit==1.39.0

langchain==0.3.6
langchain-core==0.3.15
langchain-community==0.3.6
langchain-text-splitters==0.3.0

langchain-groq==0.2.0
groq==0.11.0

sentence-transformers==2.7.0
faiss-cpu==1.8.0

python-dotenv==1.0.1
PyPDF2==3.0.1
